fix(palette): bigger resizable result window, double-click maximize, no hide-while-reviewing; fix results stuck at 56px on reopen (strip)
Split App.tsx (window sizing -> useWindowChrome) and lib.rs (disk IO -> persistence) to satisfy the 500-line monolith cap.
